Rename ConcatenatedLoRALayer to MergedLayerPatch. And other minor cleanup.
